refactor: 리뷰 수정사항 적용
