+++ client/src/routes/RestrictedRoute.tsx
@@ -1,5 +1,6 @@
 import { ReactElement } from 'react';
 import { Navigate } from 'react-router-dom';
+import { RoutePath } from '../constants';
 import { useAuthorization } from '../hooks/useAuthorization';
 
 interface Props {
@@ -11,5 +12,5 @@ export default function RestrictedRoute({ component: RouteComponent }: Props): R
   const isAuthenticated = useAuthorization();
 
   if (isAuthenticated === undefined) return <></>;
-  return isAuthenticated === true ? <Navigate to="/main/log" /> : <RouteComponent />;
+  return isAuthenticated === true ? <Navigate to={RoutePath.LOG} /> : <RouteComponent />;
 }
